Fix int32 offset overflow in Triton SituGLU (root cause of 8-node profiling faults)
With allgather_reducescatter EP, the MoE apply sees DP-gathered batches:
profiling at max-num-batched-tokens=4096 becomes M=32768 (x16 top-k rows
= 524288). _situ_and_mul_kernel computed row*2*D in int32, overflowing
past row 349,526 -> wrapped negative offsets -> GPU memory access fault.
Matches attempts 8/9 faulting at the first MoE layer during eager
profiling while the native baseline (int64 indexing) and small-M
reproducers never faulted.

Fix: int64 row index in situ kernel; same defensive cast in the Triton
rms_norm/fused_add_rms_norm kernels. concat_and_cache_mla and
mxfp4_masked_dequant audited safe (int64 slot mapping / weight-bounded).

+++ vllm/kernels/triton/rms_norm.py
@@ -47,7 +47,7 @@ def _rms_norm_kernel(
     N: tl.constexpr,
     BLOCK: tl.constexpr,
 ):
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     cols = tl.arange(0, BLOCK)
     mask = cols < N
     x = tl.load(x_ptr + row * N + cols, mask=mask, other=0.0).to(tl.float32)
@@ -71,7 +71,7 @@ def _fused_add_rms_norm_kernel(
     BLOCK: tl.constexpr,
 ):
     # In-place: y aliases x, res_ptr holds the residual in and out.
-    row = tl.program_id(0)
+    row = tl.program_id(0).to(tl.int64)
     cols = tl.arange(0, BLOCK)
     mask = cols < N
     x = tl.load(x_ptr + row * N + cols, mask=mask, other=0.0).to(tl.float32)
